+++ NEWS.md
@@ -2,6 +2,8 @@ emmeans 1.3.0.09999x
 -------------
 
   * `"mvt"` adjustment ignored `by` grouping
+  * `contrast()` mis-labeled estimates when levels varied among `by` groups
+    (most prominently this happened in `CLD(..., details = TRUE)`)
 
 
 emmeans 1.3.0