+++ NEWS.md
@@ -12,6 +12,7 @@ emmeans 1.3.0.09999x
   * Fixed bug in `hpd.summary()` and handoff to it from `summary()`
   * Fixed bug where `ref_grid()` ignored `mult.levs`
   * Fixes in emmeans where it passes `...` where it shouldn't
+  * `CLD()` now works for MCMC models (uses frequentist summary)
 
 
 emmeans 1.3.0